+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/ios/ios_lint_localizations.rb
@@ -17,8 +17,8 @@ module Fastlane
       def self.run_linter(params)
         UI.message 'Linting localizations for parameter placeholders consistency...'
 
-        require_relative '../../helper/ios/ios_l10n_helper'
-        helper = Fastlane::Helper::Ios::L10nHelper.new(
+        require_relative '../../helper/ios/ios_l10n_linter_helper'
+        helper = Fastlane::Helper::Ios::L10nLinterHelper.new(
           install_path: resolve_path(params[:install_path]),
           version: params[:version]
         )
@@ -92,7 +92,7 @@ module Fastlane
             description: 'The path where to install the SwiftGen tooling needed to run the linting process. If a relative path, should be relative to your repo_root',
             type: String,
             optional: true,
-            default_value: "vendor/swiftgen/#{Fastlane::Helper::Ios::L10nHelper::SWIFTGEN_VERSION}"
+            default_value: "vendor/swiftgen/#{Fastlane::Helper::Ios::L10nLinterHelper::SWIFTGEN_VERSION}"
           ),
           FastlaneCore::ConfigItem.new(
             key: :version,
@@ -100,7 +100,7 @@ module Fastlane
             description: 'The version of SwiftGen to install and use for linting',
             type: String,
             optional: true,
-            default_value: Fastlane::Helper::Ios::L10nHelper::SWIFTGEN_VERSION
+            default_value: Fastlane::Helper::Ios::L10nLinterHelper::SWIFTGEN_VERSION
           ),
           FastlaneCore::ConfigItem.new(
             key: :input_dir,
@@ -115,7 +115,7 @@ module Fastlane
             description: 'The language that should be used as the base language that every other language will be compared against',
             type: String,
             optional: true,
-            default_value: Fastlane::Helper::Ios::L10nHelper::DEFAULT_BASE_LANG
+            default_value: Fastlane::Helper::Ios::L10nLinterHelper::DEFAULT_BASE_LANG
           ),
           FastlaneCore::ConfigItem.new(
             key: :only_langs,